Move part of the code from ui/itemviews/columns.py to ui/columns.py
- this code can be re-used elsewhere
- code specific to itemviews stay where it was
- change imports for tests
